#42: Fix incompatible mixin in navigation CSS
`border-radius` mixin is no longer supported by Bootstrap as of
[TWBS Pull 6342](https://github.com/twbs/bootstrap/pull/6342):

>Remove .border-radius() and .border-radius mixins. As only Android
2.1, iOS 3.2, and older desktop browsers require a prefixed version,
we've removed the base mixin. Since we no longer require prefixes for
independent corners, we've dropped those mixins as well. Mixins for a
single side, like .border-left-radius, are still available.

So I just deleted the references in components/navigations.css.scss and
added that file back to the import list in application.css.scss