Restore videos tab
The cucumber specs are running in ~10s on my machine (much of which is
Rails boot time). The youtube hit is a noticable lag, but it looks like
we shoud have caching around that in production as well.

As best I can tell, this is working; am I missing anything?